refactor: 사전 식물 정보영역 컴포넌트로 추출
